Integracion del carrito con varios usuarios

+++ eliminarcarrito.php
@@ -1,5 +1,7 @@
 <?php
-    
+    require_once 'Helpers.php';
+    require_once 'Parameters.php';
+    session_start();
     $servidor='localhost';
     $cuenta='root';
     $password='';
@@ -22,18 +24,17 @@
             while( $fila = $resultado ->  fetch_assoc()){
                 $nombre = $fila['Nombre_Producto'];
                 $cantidad = $fila['Cantidad'];
-                if($nombre == $eliminar){
+                $id = $fila['Id_Usuario'];
+                if($nombre == $eliminar  && $id == $_SESSION['User']->Id){
                     if($cantidad==1){
                         $sql="DELETE FROM carrito WHERE `carrito`.`Nombre_Producto` = '$nombre'";
                     }else{
                         $cantidad--;
                         $sql = "UPDATE carrito SET Cantidad='$cantidad' WHERE Nombre_Producto='$nombre'";
                     }
-
                 }
 
             }
-
 
 
         $conexion->query($sql);  //aplicamos sentencia que inserta datos en la tabla usuarios de la base de datos
